Add new background, fix camera edges, rename ShipHandler
