+++ Assets/Prefabs/Ship Spawner.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 249fde8f5c1c3ef409e21b6972f6ef8f
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
