+++ Assets/Textures/MainMenu/Materials/space_bg.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6642eefd59726a64584edf775bab1ec1
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
